Clang-format modified files ..

+++ test_common/KokkosKernels_TestUtils.hpp
@@ -462,9 +462,11 @@ KOKKOS_INLINE_FUNCTION std::string value_type_name<Kokkos::complex<double>>() {
 int string_compare_no_case(const char* str1, const char* str2) {
   std::string str1_s(str1);
   std::string str2_s(str2);
-  for(int i=0; i<str1_s.size(); i++) str1_s[i] = std::tolower(str1_s[i]);
-  for(int i=0; i<str2_s.size(); i++) str2_s[i] = std::tolower(str2_s[i]);
-  return strcmp(str1_s.c_str(),str2_s.c_str());
+  for (size_t i = 0; i < str1_s.size(); i++)
+    str1_s[i] = std::tolower(str1_s[i]);
+  for (size_t i = 0; i < str2_s.size(); i++)
+    str2_s[i] = std::tolower(str2_s[i]);
+  return strcmp(str1_s.c_str(), str2_s.c_str());
 }
 
 }  // namespace Test
